Create a new `utils` sub-package and move the dict sorting classes there
